fix stats saving error
Musician gig totals for the season are now being saved correctly on the master orchestra roster file (vectors were previously being incorrectly passed by value to functions). When user is prompted to select a musician from a list, it is no longer possible to add a musician not on the list (previously, a musician already labelled "unavailable" could be selected by entering their vector index and the program's error checking did not catch this). Minor changes to eliminate redundant code and improve readability.